# docs/swagger.yaml
basePath: /v1
consumes:
- application/json
- text/plain
...
info:
  description: This documentation describes the Pipeliner API.
  license:
    name: MIT
    url: http://opensource.org/licenses/MIT
  title: Pipeliner API.
  version: 1.0.0
paths:
...
  /pipelines/{id}:
    get:
      operationId: getPipeline
      parameters:
      - description: id of the pipeline
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          $ref: '#/responses/Pipeline'
      summary: Get pipeline
      tags:
      - pipeline
...
produces:
- application/json
- text/html
schemes:
- http
- https
security:
- BasicAuth: []
- Token: []
- AccessToken: []
- AuthorizationHeaderToken: []
- SudoParam: []
- SudoHeader: []
- TOTPHeader: []
securityDefinitions:
  AccessToken:
    in: query
    name: access_token
    type: apiKey
  AuthorizationHeaderToken:
    description: API tokens must be prepended with "token" followed by a space.
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
  SudoHeader:
    description: Sudo API request as the user provided as the key. Admin privileges
      are required.
    in: header
    name: Sudo
    type: apiKey
  SudoParam:
    description: Sudo API request as the user provided as the key. Admin privileges
      are required.
    in: query
    name: sudo
    type: apiKey
  TOTPHeader:
    description: Must be used in combination with BasicAuth if two-factor authentication
      is enabled.
    in: header
    name: X-GITEA-OTP
    type: apiKey
  Token:
    in: query
    name: token
    type: apiKey
swagger: "2.0"